+++ packages/generate/code-builder/_functions.ts
@@ -1,6 +1,24 @@
 import { getImports } from "./_internal.ts";
 import type { Code, Import } from "./types.ts";
 
+/**
+ * Wrap an imported function in a function that generates code for
+ * a call to that function.
+ *
+ * @example
+ * ```ts
+ * const handleStuff = asFn(importNamed(
+ *   "@scope/stuff",
+ *   "handleStuff",
+ * ));
+ *
+ * const code = handleStuff("foo");
+ * ```
+ *
+ * @param imp the import of the function
+ * @param params the params of the function call
+ * @returns a code generating function
+ */
 export const asFn =
   <P extends unknown[]>(imp: Import) => (...params: P): Code => ({
     imports: [...imp.imports, ...params.flatMap(getImports)],