+++ CHANGELOG.md
@@ -12,6 +12,11 @@ This changelog will need to be split between individual packages
 
 ## [0.22.0]
 
+### Changed
+
+- [@http/generate] BREAKING: remove dependency on ts-poet, new code-builder api,
+  simplified handler generators, they must use new code-builder api
+
 ### Added
 
 - [@http/host-bun-local] new package for Bun localhost dev